% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-api.R
\name{crancache_clean}
\alias{crancache_clean}
\title{Remove the cache completely}
\usage{
crancache_clean(force = FALSE)
}
\arguments{
\item{force}{logical.  Should permissions be changed (if possible) to
    allow the file or directory to be removed?}
}
\description{
It deletes the cache directory as well.
}
\seealso{
Other cache management functions: \code{\link{crancache_list}},
  \code{\link{crancache_remove}}
}
\concept{cache management functions}
