% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-api.R
\name{crancache_list}
\alias{crancache_list}
\title{List packages in the cache}
\usage{
crancache_list()
}
\value{
A list of data frames. There is one data frame for each package
  directory. I.e. on platforms where CRAN supports binary packages,
  there is one directory for binaries, and one for source packages.
  The name of the list element is the absolute path to the package
  directory.

  Each data frame has three columns: `Package`, `Version` and `MD5sum`.
}
\description{
List packages in the cache
}
\seealso{
Other cache management functions: \code{\link{crancache_clean}},
  \code{\link{crancache_remove}}
}
