% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-api.R
\name{crancache_remove}
\alias{crancache_remove}
\title{Remove some packages from the cache}
\usage{
crancache_remove(pkgs)
}
\arguments{
\item{pkgs}{A character vector of regular expressions, that are
matched to file names in all package directories. The matching
files are removed from the cache.}
}
\description{
The function ignores errors, so if it fails to remove a
package, it just continues with the rest of the packages.
}
\seealso{
Other cache management functions: \code{\link{crancache_clean}},
  \code{\link{crancache_list}}
}
\concept{cache management functions}
