% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-packages.R
\name{download.packages}
\alias{download.packages}
\title{Download Package from CRAN-like Repositories, with Caching}
\usage{
download.packages(pkgs, destdir, available = NULL,
  repos = getOption("repos"), contriburl = contrib.url(repos, type), method,
  type = getOption("pkgType"), ...)
}
\arguments{
\item{pkgs}{
    character vector of the names of packages whose latest available
    versions should be downloaded from the repositories.
  }

\item{destdir}{
    directory where downloaded packages are to be stored.
  }

\item{available}{
    an object as returned by \code{\link{available.packages}}
    listing packages available at the repositories, or \code{NULL} which
    makes an internal call to \code{available.packages}.
  }

\item{repos}{
    character vector, the base URL(s) of the repositories
    to use, i.e., the URL of the CRAN master such as
    \code{"https://cran.r-project.org"} or its Statlib mirror,
    \code{"http://lib.stat.cmu.edu/R/CRAN"}.
  }

\item{contriburl}{
    URL(s) of the contrib sections of the
    repositories.  Use this argument only if your repository mirror is
    incomplete, e.g., because you burned only the \file{contrib} section on a
    CD.  Overrides argument \code{repos}.
  }

\item{method}{
    Download method, see \code{\link{download.file}}.
  }

\item{type}{
    character string, indicate which type of packages: see
    \code{\link{install.packages}}.
  }

\item{...}{
    additional arguments to be passed to
    \code{\link{download.file}}.
  }
}
\description{
Download Package from CRAN-like Repositories, with Caching
}
\seealso{
Other caching package manager functions: \code{\link{install.packages}}
}
