% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_files}
\alias{download_files}
\title{Download multiple files, in parallel}
\usage{
download_files(downloads)
}
\arguments{
\item{downloads}{Named list. The names are the paths for the target
files. Each list element is a character vector of URLs to try.}
}
\description{
For each file, a list of URLs can be given, and they are tried one
by one.
}
\details{
If a \code{<filename>.etag} file exists, then it is used to check if the
download is needed, with a HTTP HEAD request.
}
\keyword{internal}
