% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-dir.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Root of the cache directory}
\usage{
get_cache_dir()
}
\value{
Path to the root of the cache directory,
determined using \code{rappdirs}. E.g. on macOS it is
\code{/Users/gaborcsardi/Library/Caches/R-crancache}.
}
\description{
Root of the cache directory
}
