% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow-repos.R
\name{get_current_shadow_repos}
\alias{get_current_shadow_repos}
\title{Return current shadow repositories}
\usage{
get_current_shadow_repos(repos)
}
\arguments{
\item{repos}{Character vector of repositories. Typically just the
value of \code{getOption("repos")}.}
}
\value{
Character vector of shadow repositories.
}
\description{
Note that there is no \code{type} argument here, and we handle both binary
and source packages. We update both for each shadow, just in case
\code{utils::available_packages} needs either or both.
}
\keyword{internal}
