% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages.R
\name{install.packages}
\alias{install.packages}
\title{Install Packages from Repositories or Local Files, with Caching}
\usage{
install.packages(pkgs, lib, repos = getOption("repos"),
  contriburl = contrib.url(repos, type), method, available = NULL,
  destdir = NULL, dependencies = NA, type = getOption("pkgType"), ...)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose
   current versions should be downloaded from the repositories.

   If \code{repos = NULL}, a character vector of file paths.
   These can be source directories or archives
   or binary package archive files (as created by \command{R CMD build
     --binary}).  (\code{http://} and \code{file://} URLs are also
   accepted and the files will be downloaded and installed from local
   copies.)  On a CRAN build of \R for macOS these can be \file{.tgz}
   files containing binary package archives.
   Tilde-expansion will be done on file paths.

   If this is missing or a zero-length character vector, a listbox of
   available packages is presented where possible in an interactive \R
   session.}

\item{lib}{
    character vector giving the library directories where to
    install the packages.  Recycled as needed.  If missing, defaults to
    the first element of \code{\link{.libPaths}()}.
  }

\item{repos}{
    character vector, the base URL(s) of the repositories
    to use, e.g., the URL of a CRAN mirror such as
    \code{"https://cloud.r-project.org"}.  For more details on
    supported URL schemes see \code{\link{url}}.

    Can be \code{NULL} to install from local files, directories or URLs:
    this will be inferred by extension from \code{pkgs} if of length one.
  }

\item{contriburl}{
    URL(s) of the contrib sections of the repositories.  Use this
    argument if your repository mirror is incomplete, e.g., because
    you burned only the \file{contrib} section on a CD, or only have
    binary packages.  Overrides argument \code{repos}.
    Incompatible with \code{type = "both"}.
  }

\item{method}{
    download method, see \code{\link{download.file}}.  Unused if
    a non-\code{NULL} \code{available} is supplied.
  }

\item{available}{
    a matrix as returned by \code{\link{available.packages}}
    listing packages available at the repositories, or \code{NULL} when
    the function makes an internal call to \code{available.packages}.
    Incompatible with \code{type = "both"}.
  }

\item{destdir}{
    directory where downloaded packages are stored.  If it is
    \code{NULL} (the default) a subdirectory
    \code{downloaded_packages} of the session temporary
    directory will be used (and the files will be deleted
    at the end of the session).
  }

\item{dependencies}{logical indicating whether to also install
    uninstalled packages which these packages depend on/link
    to/import/suggest (and so on recursively).  Not used if \code{repos
    = NULL}.  Can also be a character vector, a subset of
    \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.

    Only supported if \code{lib} is of length one (or missing),
    so it is unambiguous where to install the dependent packages.  If
    this is not the case it is ignored, with a warning.

    The default, \code{NA}, means
    \code{c("Depends", "Imports", "LinkingTo")}.

    \code{TRUE} means to use
    \code{c("Depends", "Imports", "LinkingTo", "Suggests")} for
    \code{pkgs} and
    \code{c("Depends", "Imports", "LinkingTo")} for added dependencies:
    this installs all the packages needed to run \code{pkgs}, their
    examples, tests and vignettes (if the package author specified them
    correctly).

    In all of these, \code{"LinkingTo"} is omitted for binary packages.
  }

\item{type}{character, indicating the type of package to download and
    install.  Will be \code{"source"} except on Windows and some macOS
    builds: see the section on \sQuote{Binary packages} for those.
  }

\item{...}{
    Arguments to be passed to \code{\link{download.file}} or to the
    functions for binary installs on macOS and Windows (which accept
    an argument \code{"lock"}: see the section on \sQuote{Locking}).
  }
}
\description{
Install Packages from Repositories or Local Files, with Caching
}
