% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-cache.R
\name{update_cache}
\alias{update_cache}
\title{Update the cache after a download/install/update}
\usage{
update_cache(destdir, binaries = FALSE, warnings = list(), lib, timestamp,
  args)
}
\arguments{
\item{destdir}{Directory of the downloaded packages. See
\code{\link[utils:install.packages]{utils::install.packages()}}.}

\item{binaries}{Whether to try to add binary packages to the cache.}

\item{warnings}{List of warnings we got from \code{\link[utils:install.packages]{utils::install.packages()}}.}

\item{lib}{The \code{lib} argument to \code{install_packages}. Where the
packages are installed. If missing or \code{NULL}, then first element
of \code{\link[base:.libPaths]{base::.libPaths()}} is used.}

\item{timestamp}{A timestamp, when the installation started.
We use this to decide if a package was installed in the current
process.}

\item{args}{additional arguments to \code{install_packages} (or the other
functions), they have to be matched.}
}
\description{
Update the cache after a download/install/update
}
\keyword{internal}
