% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-packages.R
\name{update_packages}
\alias{update_packages}
\title{Update Outdated Packages, with Caching}
\usage{
update_packages(lib.loc = NULL, repos = getOption("repos"),
  contriburl = contrib.url(repos, type), method, instlib = NULL,
  ask = TRUE, available = NULL, oldPkgs = NULL, ..., checkBuilt = FALSE,
  type = getOption("pkgType"))
}
\arguments{
\item{lib.loc}{character vector describing the location of R
    library trees to search through (and update packages therein), or
    \code{NULL} for all known trees (see \code{\link{.libPaths}}).}

\item{repos}{
    character vector, the base URL(s) of the repositories
    to use, e.g., the URL of a CRAN mirror such as
    \code{"https://cloud.r-project.org"}.
  }

\item{contriburl}{URL(s) of the contrib sections of the
    repositories.  Use this argument if your repository is
    incomplete.  Overrides argument \code{repos}.
    Incompatible with \code{type = "both"}.
  }

\item{method}{Download method, see \code{\link{download.file}}.
    Unused if a non-\code{NULL} \code{available} is supplied.}

\item{instlib}{character string giving the library directory where to
    install the packages.}

\item{ask}{logical indicating whether to ask the user to select
    packages before they are downloaded and installed, or the character
    string \code{"graphics"}, which brings up a widget to allow the user
    to (de-)select from the list of packages which could be updated.
    (The latter value only works on systems with a GUI version of
    \code{\link{select.list}}, and is otherwise equivalent to
    \code{ask = TRUE}.)
  }

\item{available}{
    an object as returned by \code{\link{available.packages}}
    listing packages available at the repositories, or \code{NULL} which
    makes an internal call to \code{available.packages}.
    Incompatible with \code{type = "both"}.
  }

\item{oldPkgs}{if specified as non-NULL, \code{update_packages()}
only considers these packages for updating. This may be a character
vector of package names or a matrix as returned by \code{old.packages}.}

\item{...}{additional arguments are passed to
\code{\link[utils:update.packages]{utils::update.packages()}}.}

\item{checkBuilt}{If \code{TRUE}, a package built under an earlier
    major.minor version of \R (e.g., \code{3.1}) is considered to be
    \sQuote{old}.}

\item{type}{character, indicating the type of package to download and
    install.  See \code{\link{install.packages}}.
  }
}
\description{
Update Outdated Packages, with Caching
}
\seealso{
Other caching package manager functions: \code{\link{download_packages}},
  \code{\link{install_packages}}
}
