% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{crancache}
\alias{crancache}
\alias{crancache-package}
\title{Transparent Caching of Packages from CRAN-like Repositories}
\description{
Provides a wrapper for 'install.packages()', that transparently caches
downloaded packages in a local CRAN-like repository.
}
