% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-api.R
\name{crancache_list}
\alias{crancache_list}
\title{List packages in the cache}
\usage{
crancache_list()
}
\value{
A data frames. Each data frame has four columns:
\code{Package}, \code{Repository}, \code{Version} and \code{MD5sum}.
}
\description{
List packages in the cache
}
\seealso{
Other cache management functions: \code{\link{crancache_clean}},
  \code{\link{crancache_remove}}
}
